L3 auto and an auto docs added
