+++ 2025REDOBuildSeason/src/main/java/frc/robot/Auto/AutoMissionChooser.java
@@ -1,6 +1,11 @@
 package frc.robot.Auto;
 
 import frc.robot.Auto.Missions.*;
+import frc.robot.Auto.Missions.BlueMissions.BlueScoreL2;
+import frc.robot.Auto.Missions.BlueMissions.BlueScoreL4;
+import frc.robot.Auto.Missions.RedMissions.RedScoreL2;
+import frc.robot.Auto.Missions.RedMissions.RedScoreL4;
+
 import java.util.Optional;
 
 import edu.wpi.first.wpilibj.DriverStation;
