+++ 2025REDOBuildSeason/src/main/java/frc/robot/Auto/Missions/RedMissions/RedScoreL2.java
@@ -1,8 +1,9 @@
-package frc.robot.Auto.Missions;
+package frc.robot.Auto.Missions.RedMissions;
 
 
 import frc.robot.Auto.AutoMissionChooser;
 import frc.robot.Auto.AutoMissionEndedException;
+import frc.robot.Auto.Missions.MissionBase;
 
 import frc.robot.Auto.Actions.DriveForTimeAction;
 import frc.robot.Auto.Actions.EffectorAction;
@@ -12,21 +13,21 @@ import frc.robot.Auto.Actions.WaitAction;
 
 /*
  * This mission has the robot move out of the robot starting
- * zone to score a coral on L4
+ * zone to score a coral on L2
  */
 
-public class RedScoreL4 extends MissionBase{
+public class RedScoreL2 extends MissionBase{
     @Override
     protected void routine() throws AutoMissionEndedException {
        
         //runAction(new WaitAction(AutoMissionChooser.delay)); // MAY NOT BE NEEDED
         //runAction(new DriveForTimeAction(-0.5,  4));
-        //runAction(new TurnDegreesAction(30, 1));// add 5 degrees because michalangelo has consistently com up 5 degrees short of the needed turn
+        //runAction(new TurnDegreesAction(65, 1));// add 5 degrees because michalangelo has consistently com up 5 degrees short of the needed turn
         
         // String options: passive, ramp, Score L1, Score L2, Score L3, Score L4, ejecting coral
-        runAction(new MoveElevatorAction(2, "Score L4")); // May be needed depending on how robot starts
+        runAction(new MoveElevatorAction(1, "Score L2"));
         runAction(new EffectorAction(0.12,1));
-        runAction(new DriveForTimeAction(0.5, 0.2));
         runAction(new MoveElevatorAction(2, "ramp"));
+
     }
 }
